requirements.txt
django-admin
requests
beautifulsoup4